move gcshelper out of component_builder
